1_1

+++ NeuralNetwork/src/lib/InputLayer.hpp
@@ -38,7 +38,7 @@ private:
 	void UpdateNodeScore() override{}
 
 	void SetData(std::array<typename InputInfo_::type, InputInfo_::dim> const& input){
-		for (uint i = 0, end = NodeNum(); i < end; ++i) (*this)[i]->Score(input[i]);
+		for (size_t i = 0, end = NodeNum(); i < end; ++i) (*this)[i]->Score(input[i]);
 	}
 
 public:
